requirements.txt
absl-py==2.2.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
filelock==3.16.1
fsspec==2025.3.0
google-auth==2.40.1
google-auth-oauthlib==1.0.0
grpcio==1.70.0
huggingface-hub==0.31.2
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.6
Markdown==3.7
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
packaging==25.0
pillow==10.4.0
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
runner==0.0.1
safetensors==0.5.3
si==0.1.1.post1
six==1.17.0
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tokenizers==0.20.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
typing_extensions==4.13.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
